+++ packages/orq-rc/RELEASES.md
@@ -4819,3 +4819,13 @@ Based on:
 - [python v4.2.0-rc.44] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.44] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.44 - packages/orq-rc
+
+## 2026-01-20 04:26:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.689.0 (2.796.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.45] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.45] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.45 - packages/orq-rc